some updates haha idk

+++ src/components/HomePageStudentList.jsx
@@ -1,4 +1,4 @@
-import React, { useContext } from 'react';
+import React, { useContext } from "react";
 import {
   IonAvatar,
   IonContent,
@@ -15,49 +15,50 @@ import {
   IonCard,
   IonCardContent,
   IonText,
-  IonButton
-} from '@ionic/react';
-import { pencil, trash } from 'ionicons/icons';
-import {
-  deleteStudentById
-} from "../dataservice.tsx";
-import { ScannerContext } from '../services/ScannerContext.jsx'
+  IonButton,
+} from "@ionic/react";
+import { pencil, trash } from "ionicons/icons";
+import { deleteStudentById } from "../dataservice.tsx";
+import { ScannerContext } from "../services/ScannerContext.jsx";
 
 function HomePageStudentList({ name, id, id_number }) {
-  const { recorded, setRecorded } = useContext(ScannerContext)
+  const { recorded, setRecorded } = useContext(ScannerContext);
 
-  const deleteStudent = async(id) => {
+  const deleteStudent = async (id) => {
     try {
-      await deleteStudentById(id)
-      setRecorded(!recorded)
-    } catch(err) {
-      alert('Error deleting: ' + err)
+      if(alert('Are you sure you want to delete? records will be lost and cannot be recovered.')) {
+        await deleteStudentById(id);
+      }
+      setRecorded(!recorded);
+    } catch (err) {
+      alert("Error deleting: " + err);
     }
-  }
+  };
 
   return (
     <>
-    <IonCard>
-      <IonCardContent style={{ display: 'flex', flexDirection: 'row', justifyContent: 'between', alignItems: 'center' }}>
-        <div style={{ flexGrow: 1 }}>
-          <IonText color='dark'>
-            <span style={{ fontWeight: 'bold' }}>{ name }</span>
-          </IonText>
-            <sm style={{ fontSize: '.9em'}}>{ id_number }</sm>
-        </div>
-        <div style={{ display: 'flex', flexDirection: 'row', gap: '6px'}}>
-          <IonButton color='success' class='edit-btn' routerLink={`edit/${id}`}>
-            <IonIcon icon={pencil} />
-          </IonButton>
-          <IonButton color='danger' class='trash-btn' onClick={() => deleteStudent(id)}>
-            <IonIcon  icon={trash} />
-          </IonButton>
-        </div>
-      </IonCardContent>
-    </IonCard>
+      <IonCard>
+        <IonCardContent
+          style={{ display: "flex", justifyContent: "space-between" }}
+        >
+          <div>
+            <IonText color="dark">
+              <span style={{ fontWeight: "bold" }}>{name}</span>
+            </IonText>
+            <sm style={{ fontSize: ".9em" }}>{id_number}</sm>
+          </div>
+          <div style={{ display: "flex", flexDirection: "row", gap: "6px" }}>
+            <IonButton fill='outline'  shape='round' routerLink={`edit/${id}`}>
+              <IonIcon slot='icon-only' icon={pencil} />
+            </IonButton>
+            <IonButton fill='outline' shape='round' onClick={() => deleteStudent(id)}>
+              <IonIcon slot='icon-only' icon={trash} />
+            </IonButton>
+          </div>
+        </IonCardContent>
+      </IonCard>
 
-
-    {/*  <IonItemSliding>
+      {/*  <IonItemSliding>
             <IonItem button={true}>
               <IonLabel color='dark'>{ name }</IonLabel>
             </IonItem>
